requirements.txt
curio==1.4
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==7.2.0
